Refactor : AiChatCountService 인터페이스 분리 및 구현 클래스 추가
